BAH-1544 | refactor. paginated table as a separate component
