+++ config/spa-build-config.json
@@ -1,10 +1,8 @@
 {
     "publicUrl": "/lab",
     "frontendModules": {
-      "@openmrs/esm-patient-search-app":"3.1.0",
       "@openmrs/esm-patient-banner-app":"3.1.0",
       "@openmrs/esm-patient-registration-app" :"3.1.0",
-      "@openmrs/esm-active-visits-app" :"3.1.0",
-      "@openmrs/esm-home-app": "3.1.1"
+      "@openmrs/esm-active-visits-app" :"3.1.0"
     }
 }
