+++ src/constants.ts
@@ -4,6 +4,8 @@ export const searchModuleName = '@openmrs/esm-patient-search-app'
 export const patientLabDetailsPath = 'patient/${patientUuid}'
 export const patientLabDetailsRoute = '/patient/:patientUuid'
 export const labOrderUuid = '8189b409-3f10-11e4-adec-0800271c1b75'
+export const defaultPageSize = 5
+
 
 export const headers = [
     {
